260208E: MathUtils multi-type templates + dead code cleanup
- clamp(T,T,T) → clamp(T1,T2,T3) with common_type return
- map(T,T,T,T,T) → map(T1..T5) returning float (fixes integer division)
- mapRange() → template accepting mixed types
- Removed ~22 static_cast<float> wrappers across 6 files
- Deleted dead LightConduct.cpp, removed setMeasurementEnabled API
- brightnessFading moved to Globals (cross-lib visibility, no extern)
- PatternCatalog/ColorsCatalog next/prev: PF → LOG_DEBUG

+++ lib/AudioManager/AudioState.cpp
@@ -49,9 +49,8 @@ int getAudioSliderPct() {
     float effectiveHi = shiftedHi * webShift;
     // Map to slider percentage using Globals (like brightness)
     return static_cast<int>(MathUtils::mapRange(
-        effectiveHi,
-        Globals::volumeLo, Globals::volumeHi,
-        static_cast<float>(Globals::loPct), static_cast<float>(Globals::hiPct)));
+        effectiveHi, Globals::volumeLo, Globals::volumeHi,
+        Globals::loPct, Globals::hiPct));
 }
 
 void setAudioLevelRaw(int16_t value) {
